глава 8 задача 01 Перевел с Int на Double

+++ src/main/scala/chapter8/ex01.scala
@@ -7,14 +7,14 @@ package chapter8 {
   import chapter5.ex02.BankAccount
 
   object ex01 extends App {
-    class CheckingAccount(private val current: Int = 0) extends BankAccount(current) {
-      def service(v: Int) = v-1
-      override def deposit (v: Int): CheckingAccount = new CheckingAccount(service(super.deposit(v).balance))
-      override def withdraw(v: Int): CheckingAccount = new CheckingAccount(service(super.withdraw(v).balance))
+    class CheckingAccount(private val current: Double = 0) extends BankAccount(current) {
+      def service(v: Double) = v-1
+      override def deposit (v: Double): CheckingAccount = new CheckingAccount(service(super.deposit(v).balance))
+      override def withdraw(v: Double): CheckingAccount = new CheckingAccount(service(super.withdraw(v).balance))
     }
 
     object CheckingAccount {
-      def apply(v: Int): CheckingAccount = new CheckingAccount(v)
+      def apply(v: Double): CheckingAccount = new CheckingAccount(v)
     }
 
     assert(CheckingAccount(100).deposit(30).withdraw(15).balance == 113)
